requirements/default.txt
djangorestframework>=2.4.3
django-oauth-toolkit>=0.10.0
django-otp>=0.3.5
six>=1.10.0
twilio>=6.0.0
Django>=4.0,<4.1
